LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/error-type-1-expected.txt


FAIL network error has higher priority than parse error assert_equals: expected 3 but got 2
